chromium: remove patches already applied by meta-imx
The chromium-ozone-wayland patches were introduced to enable support for
chromium on i.MX platform. Those patches derived from NXP's meta-imx.
Now that the meta-imx is included in the release, the same patches are
applied twice, resulting in errors during the build. Remove the patches
from this meta-layer, since they are already in meta-imx.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>