var-debian-xwayland: use meta-nxp-desktop graphics
The initial intention was to remove as much depenendcies from
meta-nxp-desktop as possible, as it is not clear how long this
layer will be support by NXP.

But without we're expierencing problems with the graphics. As
a quick solution, let's consider graphics from the desktop
layer for now.

The prefered version for imx-gpu-viv will be removed because of
some issues related to it.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>