variscite: scripts: mx6ul_mx7_install_yocto: add support for codec selection
Beginning with DART-6UL v1.3, respectivaly DART-6UL-5G v2.1 the codec wm8731
has been replaced with the codec wm8904.

For NAND installation we can only install a single device tree. Add support
to select the right one based on what can be detected at runtime.

The runtime selection relays on trying to detect the codec using the i2c bus.

If there is no codec at the expected address, it will fallback to the default
device tree.

If there is a codec, and it's the legacy one (wm8731), it will consider using
the device tree supporting this codec. When it's the new codec, just use the
default device tree name.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>