variscite: scripts: mx6ul_mx7_install_yocto: add support for codec selection on VAR-SOM-MX7
This commit is based on the following commit:

varigit/meta-variscite-sdk-imx@0cdcfa062d27 variscite: scripts: mx6ul_mx7_install_yocto: add support for codec selection

The codec wm8731 has been replaced with the codec wm8904, for NAND installation
it is only possible to install a single device tree. Add support to select the
right one based on what can be detected at runtime.

The runtime selection relays on trying to detect the codec using the i2c bus. If
there is no codec at the expected address, it will fallback to the default device
tree. If there is a codec, and it is the legacy one (wm8731), it will consider
using the device tree supporting this codec. When it is the new codec, just use
the default device tree name.

Signed-off-by: Diego Dorta <diego.d@variscite.com>