swupdate: defconfig: update to swupdate-2024.12
This is the result of doing `make menuconfig` with the previous
defconfig and exiting directly. No changes were made in the menu.

It is not allowed to do a `savedefconfig` to create a reduced version of
the config as the complete file and potentially existing config
fragments are parsed from the Yocto recipe to determine dependencies.

Also, extra compiler flags were manually removed as they contain part of
the build environment. This has also be done upstream, see here for
example:
https://github.com/sbabic/meta-swupdate/commit/f1178606a028723dde2ebe91c43f84cff270bce0

Signed-off-by: Matthias Proske <matthias.p@variscite.com>