swupdate: update swupdate configuration files
This replaces the start of the swupdate and brings them to the current
version, also allowing future updates to be taken over automatically.
Configurations are loaded from /etc/swupdate/conf.d instead of being
passed on as command line parameters.

This solves issues with swupdate not terminating properly after an
update with a reboot (an extra 90 timeout was needed before).

Signed-off-by: Matthias Proske <matthias.p@variscite.com>