scripts/var-setup-release: Enable buildhistory
Build History in Yocto is a feature that allows tracking and analyzing
changes and details of each build (compilation) performed in a Yocto project.
It collects detailed information about packages, images, executed tasks,
dependencies, and much more, storing this data in a specific directory for later
analysis.

This feature is useful for:

1. Auditing changes: Comparing previous builds to understand what has changed between them.
2. Debugging: Identifying issues or differences between builds.
3. Quality control: Ensuring that builds are consistent and that unwanted changes are not introduced.
4. Documentation: Maintaining a history of builds for future reference.

When Build History is enabled, Yocto collects information during the build
process and stores it in a directory called `buildhistory`, located in the
build directory (usually `tmp/buildhistory`).

The collected data includes:

- Package information: Details about generated packages, such as versions, dependencies, licenses, etc.
- Images: Information about generated images (e.g., `core-image-minimal`).
- Tasks: Details about tasks executed during the build.
- Dependencies: Relationships between packages and tasks.
- Differences between builds: Comparisons between previous and current builds.

For information on the build history feature, see the
"Maintaining Build Output Quality" section in the Yocto Project Development
Tasks Manual.

Signed-off-by: Luan Rafael Carneiro <luan.r@variscite.com>