[RND-2729]: swupdate: update.sh: use TMPDIR /tmp if not given in environment
Set TMPDIR to /tmp if not already set.
According to our documentation, it is optional to set TMPDIR. If it is
not set, swupdate will use /tmp automatically, but our update.sh scripts
need it be set explicitly.
In case of mx6ul, mx7 and mx6 the swupdate process would fail if not
set.
On other platforms, the / folder is used as a temporary directory,
leaving behind a dummy folder in the original partition.

Also aligning all platforms to create the folder ${TMPDIR}/datadst which
should no longer be necessary, but allows us to be independent of any
underlying changes that swupdate might do.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>