mx6ul: Desktop icon: also provide "Install Yocto" icons when using var-recovery-image
When creating an SD card, using var-create-yocto-sdcard.sh, icons were
installed for the GUI to allow installation to eMMC / NAND directly from
the touchscreen.
Add these icons also when creating an SD card using var-recovery-image.

If dual-partitioning is used is based on the name of the image to be
installed - similar to what was used before, but now an additional
helper script is used to determine the correct parameters for the
installation.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>