xserver-nodm: allow X Server (and sub processes) to start faster during boot
With the switch to systemd, the boot processes are highly paralellized.
This causes that the start of the X server and its subprocesses are
often interrupted. While the display already turns black when the X
server starts its initialization (removing any progress bar that might
indicate the boot status), it might take a few seconds before the full
desktop (started by the subprocesses) is shown as systemd starts more
services that also consume the CPU. This leaves the user with a black
screen - for several seconds in the worst case scenario.
This adjusts the priority for the X server and its sub processes,
allowing the startup to execute quickly even when lots of other
processes also want to use the CPU. The result is that the switch from
the progress bar to the normal desktop is way better.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>