xserver-nodm: wait for udev devices to be settled
If starting the X server too early, the touchscreen is not up and
running yet.
This means that the Desktop is visible and there is even a mouse cursor,
but you cannot move that mouse cursor until the touchscreen becomes
available which can be >10 seconds on imx6ul.
Let's delay the start of the X server until the touchscreen is
available. To have a generic solution in which we don't care if we have
resistive touch or a capacitive touch (or maybe even no touch at all),
let's wait for all udev events to finish before starting the X server.
During this time the progress bar splash screen is still updated,
indicating some progress.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>