fsl-image-gui-chromium: Enable onscreen keyboard by default in Chromium
Enable the onscreen keyboard by default to improve the user experience
on our EVKs that are available with a touchscreen display. This requires
passing the `--enable-wayland-ime` option to Chromium, which activates
the Weston on-screen keyboard (`weston-keyboard`) on demand.

Currently, the weston launcher command in weston.ini is:

    path=/usr/sbin/runuser -l weston -c chromium

To preserve this call method while adding the flag, quoting or escaping
would be required:

    path=/usr/sbin/runuser -l weston -c 'chromium --enable-wayland-ime'
    or
    path=/usr/sbin/runuser -l weston -c chromium\ --enable-wayland-ime

Since weston.ini does not support this, a helper script is needed to
work around the limitation.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>
(cherry picked from commit 45a234c9d9d31593795a9232b54340286b6179b9)