libcamera: add patch to enable video streaming with resolution above 1920x1080
The patch fixes an issue affecting the imx8-isi pipeline, which is used by all imx8/imx9 SoCs.
The original code in PipelineHandlerISI::generateConfiguration() limits the camera resolution
to 1920x1080 unless the pipeline stream-role property is set to still-capture, which is not the default.
The modified code instead will limit the resolution only when stream-role is set to view-finder,
which is not the default as well, leaving it unlimited by default.

Signed-off-by: Marco Contenti <marco.c@variscite.com>
(cherry picked from commit bb40f5fc8295422a8abc35a37fa15e6715ba4157)